requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
python-dotenv
PyJWT
requests
